git-svn-id: file:///media/LinuxData/Development/SVN/LFSDesktopProject@272 3ebf34f5-1e21-47b0-85e2-b17d3fdaf10f

+++ LFSToolKit/LFSToolKit/src/LFSTKButton.cpp
@@ -150,7 +150,7 @@ listener* LFSTK_buttonClass::LFSTK_getListen(void)
 	return(&(this->listen));
 }
 
-void LFSTK_buttonClass::LFSTK_setCallBack(void (*downcb)(void *,int),void (*releasecb)(void *,int),int ud)
+void LFSTK_buttonClass::LFSTK_setCallBack(void (*downcb)(void *,int),void (*releasecb)(void *,int),long ud)
 {
 	this->callback.pressCallback=downcb;
 	this->callback.releaseCallback=releasecb;
